🔨 refactor(app): remove features folder
