+++ frontend/src/components/RequireAuth/index.ts
@@ -1,0 +1,1 @@
+export * from "./RequireAuth";
