--- frontend/src/features/auth/index.ts
@@ -1,2 +1,0 @@
-export { LoginForm } from "./LoginForm";
-export { RequireAuth } from "./RequireAuth";
